requirements.txt
apache-airflow==2.10.5
pandas
requests
selenium==4.10.0
selenium-stealth
openpyxl
matplotlib
seaborn
xgboost
scikit-learn
pmdarima
statsmodels
joblib
lightgbm